// shencha/source_code/src/types/workflow.ts
export interface SceneOption {
  id: string;
  name: string;
  priority: "P0" | "P1" | "P2";
  template_id?: string;
}

export interface PrimarySceneOption extends SceneOption {
  subScenes?: SceneOption[];
}